refactor: refactoring some functions and components to apply ssr
