+++ frontend/src/components/Headers/LoggedHeader.tsx
@@ -1,15 +1,18 @@
+'use client'
+
+import { useContextWrap } from '@/contextAPI/context'
+import checkAuth from '@/functions/checkAuthFunction'
+import useLogOut from '@/functions/logOutFunction'
 import Link from 'next/link'
-import { SetStateAction } from 'react'
 
-interface LoggedHeaderProps{
-  onClick: (event:React.MouseEvent<HTMLButtonElement>) => void
-  setNotGuest: React.Dispatch<SetStateAction<boolean>>
-}
+// interface LoggedHeaderProps{
+//   onClick: (event:React.MouseEvent<HTMLButtonElement>) => void
+// }
 
-export default function LoggedHeader({
-  onClick,
-  setNotGuest
-}:LoggedHeaderProps) {
+export default function LoggedHeader() {
+  checkAuth()
+  const {setNotGuest} = useContextWrap()
+  const logOut = useLogOut()
   return (
     <>
       <div className="py-4">
@@ -46,7 +49,7 @@ export default function LoggedHeader({
           </div>
           <div className="">
             <Link href="/">
-              <button onClick={onClick} className=" font-semibold border-solid border-gray-200 border-2 rounded-3xl px-5 py-2 mr-5 hover:bg-gray-200">
+              <button onClick={logOut} className=" font-semibold border-solid border-gray-200 border-2 rounded-3xl px-5 py-2 mr-5 hover:bg-gray-200">
                 Logout
               </button>
             </Link>
