+++ frontend/src/components/weddingsDisplay/GuestWeddings.tsx
@@ -1,12 +1,63 @@
-import Link from 'next/link'
-import weddingListProps from '@/types/weddingListProps'
+'use client'
 
-export default function WeddingsGuest({ weddingsArray }: weddingListProps) {
+import Link from 'next/link'
+import { useContextWrap } from '@/contextAPI/context'
+import { useEffect, useState } from 'react'
+import weddingProps from '@/types/weddingProps'
+import axios from 'axios'
+
+export default function WeddingsGuest() {
+  const {userToken} = useContextWrap()
+  const [guestWeddingsArray, setGuestWeddingsArray] = useState<weddingProps[]>([
+      {
+        id: 0,
+        weddingTitle: '',
+        weddingDate: '',
+        shippingAddress: '',
+        createdBy: '',
+      },
+    ])
+    useEffect(() => {
+      async function getWeddings() {
+        if (userToken)
+          try {
+            const response = await axios.get(
+              'http://localhost:3000/getWeddings',
+              {
+                headers: {
+                  Authorization: `Bearer ${userToken}`,
+                },
+              }
+            )
+            if (response.status === 200) {
+              setGuestWeddingsArray(response.data.invited)
+            }
+          } catch (error) {
+            if (axios.isAxiosError(error)) {
+              if (error.response?.status === 401) {
+                console.log('User not authenticated.')
+              }
+              if (error.response?.status === 403) {
+                console.log('Invalid/Expired token.')
+              }
+              if (error.response?.status === 404) {
+                console.log('User not found.')
+              }
+              if (error.response?.status === 500) {
+                console.log('Server error.')
+              }
+            }
+          }
+      }
+  
+      getWeddings()
+    }, [userToken])
+    
   return (
     <ul className="flex flex-col text-center items-center">
-      {weddingsArray.map((wedding) => (
+      {guestWeddingsArray.map((wedding) => (
         <div
-          id={wedding.id}
+          id={`${wedding.id}`}
           key={wedding.id}
           className="p-3 sm:p-5 border-gray-400 w-full sm:w-3/5"
         >
